Row-Major Sparse Matrices for Scalability tests

+++ famu/setup_hessian_modes.cpp
@@ -50,14 +50,14 @@ using namespace std;
 
 void famu::setup_hessian_modes(Store& store, SparseMatrix<double>& A, MatrixXd& mG){
         int nummodes = store.jinput["number_modes"];
-        nummodes = std::min(nummodes, store.S.cols());
+        nummodes = std::min(nummodes, (int)store.S.cols());
 
         cout<<"+EIG SOLVE"<<endl;
 
         cout<<"     eig1"<<endl;
         //Spectra seems to freak out if you use row storage, this copy just ensures everything is setup the way the solver likes
         Spectra::SparseSymMatProd<double>Aop(A);
-        Spectra::SymEigsSolver<double, Spectra::SMALLEST_MAGN, Spectra::SparseSymMatProd<double> > geigs(&Aop, nummodes, std::min(5*nummodes, A.rows()) );
+        Spectra::SymEigsSolver<double, Spectra::SMALLEST_MAGN, Spectra::SparseSymMatProd<double> > geigs(&Aop, nummodes, std::min(5*nummodes, (int)A.rows()) );
         geigs.init();
  
         cout<<"     eig2"<<endl;
